Split pastBills into body and footer directories
