+++ components/pastBills/body/pill.jsx
