+++ components/pastBills/footer/footer.jsx
@@ -1,0 +1,7 @@
+import React, { useState, useEffect } from 'react';
+
+export default function Footer() {
+  return (
+    <div></div>
+  );
+}
