+++ components/pastBills/footer/index.js
@@ -1,0 +1,2 @@
+import Footer from './footer.jsx';
+export default Footer;
